Improve intuitivity of available availability zones list
As suggested by @bzwei, this patch improves the way the availability
zone list is generated. In case no zones are available, the returned
list shows just <none> option, if there is exactly one, it is selected
automatically, but if there are more than one, the <select> option is
added as the default one and the user needs to choose one from the list.

Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>


(transferred from ManageIQ/manageiq@1782df1175aba7a59eb6da8f7271019e2091c4a8)