Refresh based on all saved Amazon EmsEvents
Refresh based on all saved Amazon EmsEvents in a way that
we can override the behaviour for single EventTypes
from automate.

https://github.com/ManageIQ/manageiq-providers-amazon/pull/93
needs to be reverted, since it changed all unknown event_types
to one. And The list of recognized events was in the code.

THis PR has the same behaviour, but all is configurable from
Automate.

fixes:
https://bugzilla.redhat.com/show_bug.cgi?id=1371222
https://github.com/ManageIQ/manageiq/issues/10591