Check for both owner email and requester email for user quota.
Changed used method to check for owner email or requester email when quota is set to user.
Previously the method was only looking for owner_email and when that wasn't available we encountered an error.

The method now will raise an error if both are not found.  If you are running a service provision which doesn't allow a requester email
and the user doesn't have an email address, we will raise this error:

ERROR - Owner email not specified for User Quota

Added tests for all possible email values for user quota.

https://bugzilla.redhat.com/show_bug.cgi?id=1509977