Changed Cloud VM Provision validate_request max_memory to be consistent.
The values for approval are confusing as Jeffrey Cutter mentioned in his BZ.

I changed instance values to be in Megabytes.  They were in bytes.
The instance values for quota are in megabytes so the approval values will match now.

Tag values should match the UI and be in gigabytes.  They were in Megabytes.
I changed tag values to be in Gigabytes.

I also human_sized all the log messages so going forward it will be clear what the values really are.
I will be doing another PR for Infrastructure with the same changes.

https://bugzilla.redhat.com/show_bug.cgi?id=1553530