Added quota calculations for number_of_vms in service dialogs for Infrastructure.
If you specify number_of_vms in a dialog override, the changes will multiple this value by memory, storage and number_of_cpus for quota value.
If you are requesting 1gb of memory, 1gb of storage and 1 cpu and override number_of_vms to 3,
quota will now check for 3gb of memory, 3gb of storage,  3 cpus and 3 vms.

Modified a test and added a new test.
Fixed rubocop warning.
Changed total_vms to a local variable.