Backport of PR #390
Prevent the module from running when required by hooking the class definition and stubbing the call to .new during require. Doing this ensures that the .new calls a real class that already exists.

Unfortunately, the class name is occasionally misspelled and this code path was previously not covered. This adds test coverage for this path.