Allow OrchestrationStack retirement be triggered from Service
Current stack retirement automation assumes being triggered on stack
directly, but we want to be able to use the entrypoint

```
/Cloud/Orchestration/Retirement/StateMachines/StackRetirement/Default
```

from the service that handles stack as well. With this commit we therefore
adjust existing stack retirement workflow to accept stack passed either
as

```
$evm.root['orchestration_stack']
```

or as

```
$evm.root['service'].orchestration_stack
```

In latter case we also make sure that Service status gets properly
updated (so that the service is then moved to "Retired Services"
directory).

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1632239

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>