Use the Ems last_inventory_date for ReconfigVM
Instead of using a task to wait for a refresh to finish, check the EMS's
last_inventory_date which mark the date when the inventory was
collected.

This lets us check if the inventory in the DB includes the changes which
an Event is being processed for.

This is important because streaming refresh doesn't use the typical
EmsRefresh.refresh workflow and so no task gets created.