Fix Service Retirement requests leaving the Service is a state of 'retiring'.
This will fix the problem when someone uses the older api call and leaves the service in a 'retiring' state.
If there is no task, the process is aborted and the state is 'initializing'.

This will allow the retirement process will start again instead of being denied because the state is 'retiring'

fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1698480

Updated update_service_retirement_status_spec
Moved check for service_retire_task before notification as requested.
Removed tap and changed code  as requested
Changed 'fred' to 'active' as requested