Reset ae_reason and ae_result in check_refreshed when refresh_may_have_completed is true.
The update_result and update_reason state_vars contained the
previous values.  The state_var update_result
value was still set to retry from the previous execution which caused
the state machine to retry instead of proceeding to the next state.
It had another side affect in that the retry was set with no retry
interval which caused the state machine to retry immediately instead of
waiting the usual time interval.

Resetting those values caused the state machine to progress to the next
state and complete successfully.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1720212