Adding better logging to check_service_retire method.
Now you will see the task ids, type and descriptions of tasks waiting to be retired.

Service Retire Task:<10000000000118> <Service Retire for: vmware2-20191218-122517> is not retired, setting retry.
  Service Retire Task:<10000000000118> waiting on <vm_retire> Task:<10000000000119> <VM Retire for: Sample_Vm>

Changed method name to  log_task_info as requested
Changed : with space as requested
Changed 'waiting on' to 'is waiting on' as requested